refactor(bzm2): extract hash thread hashing helpers
